+++ project_ww/app.py
@@ -12,14 +12,16 @@ from wtforms.fields.simple import StringField, PasswordField
 from wtforms.validators import DataRequired, Email, EqualTo
 from config import Config
 from flask_wtf.csrf import CSRFProtect
+from models import File,User,db
 
 app = Flask(__name__)
 app.config.from_object(Config)
+db.init_app(app)
+migrate=Migrate(app,db)
 
 # Inicjalizacja CSRFProtect
 csrf = CSRFProtect(app)
 
-db = SQLAlchemy(app)
 migrate = Migrate(app, db)
 
 login_manager = LoginManager(app)
@@ -34,24 +36,6 @@ if not os.path.exists(os.path.join(app.root_path, 'instance')):
     os.makedirs(os.path.join(app.root_path, 'instance'))
 
 # Model User
-class User(UserMixin, db.Model):
-    user_id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(50), unique=True, nullable=False)
-    email = db.Column(db.String(100), unique=True, nullable=False)
-    password = db.Column(db.String(255), nullable=False)
-
-    def get_id(self):
-        return self.user_id
-
-# Model File
-class File(db.Model):
-    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    filename = db.Column(db.String(120), unique=True, nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.user_id'), nullable=False)
-
-    def __init__(self, filename, user_id):
-        self.filename = filename
-        self.user_id = user_id
 
 @login_manager.user_loader
 def load_user(user_id):
